Move responsibility for undo and redo of tracks to a new file
