+++ libraries/lib-track/Track.cpp
@@ -27,7 +27,6 @@ and TimeTrack.
 
 #include "BasicUI.h"
 #include "Project.h"
-#include "UndoManager.h"
 
 #include "InconsistencyException.h"
 
@@ -1375,50 +1374,6 @@ void Track::OnProjectTempoChange(double newTempo)
 const std::optional<double>& Track::GetProjectTempo() const
 {
    return GetGroupData().mProjectTempo;
-}
-
-// Undo/redo handling of selection changes
-namespace {
-struct TrackListRestorer final : UndoStateExtension {
-   TrackListRestorer(AudacityProject &project)
-      : mpTracks{ TrackList::Create(nullptr) }
-   {
-      for (auto pTrack : TrackList::Get(project)) {
-         if (pTrack->GetId() == TrackId{})
-            // Don't copy a pending added track
-            continue;
-         mpTracks->Append(std::move(*pTrack->Duplicate()));
-      }
-   }
-   void RestoreUndoRedoState(AudacityProject &project) override {
-      auto &dstTracks = TrackList::Get(project);
-      dstTracks.Clear();
-      for (auto pTrack : *mpTracks)
-         dstTracks.Append(std::move(*pTrack->Duplicate()));
-   }
-   bool CanUndoOrRedo(const AudacityProject &project) override {
-      return !TrackList::Get(project).HasPendingTracks();
-   }
-   const std::shared_ptr<TrackList> mpTracks;
-};
-
-UndoRedoExtensionRegistry::Entry sEntry {
-   [](AudacityProject &project) -> std::shared_ptr<UndoStateExtension> {
-      return std::make_shared<TrackListRestorer>(project);
-   }
-};
-}
-
-TrackList *TrackList::FindUndoTracks(const UndoStackElem &state)
-{
-   auto &exts = state.state.extensions;
-   auto end = exts.end(),
-      iter = std::find_if(exts.begin(), end, [](auto &pExt){
-         return dynamic_cast<TrackListRestorer*>(pExt.get());
-      });
-   if (iter != end)
-      return static_cast<TrackListRestorer*>(iter->get())->mpTracks.get();
-   return nullptr;
 }
 
 TrackListHolder TrackList::Temporary(AudacityProject *pProject,
